+++ CHANGELOG.md
@@ -16,7 +16,8 @@ See [RELEASE](./RELEASE.md) for workflow instructions.
 
 * [#5803](https://github.com/spacemeshos/go-spacemesh/pull/5803) Fixed PoST verifiers autoscaling for 1:N setups.
 
-### Features
+* [#5819](https://github.com/spacemeshos/go-spacemesh/pull/5819) The node will now refuse connections from post services
+  if no coinbase account is set.
 
 ## Release v1.4.4